removed governance page
