+++ src/providers/Fractal/governance/types.ts
@@ -2,6 +2,10 @@ import {
   FractalModule,
   FractalUsul,
   OZLinearVoting,
+  UsulVetoGuard,
+  VetoERC20Voting,
+  VetoGuard,
+  VetoMultisigVoting,
   VotesToken,
 } from '@fractal-framework/fractal-contracts';
 import {
@@ -71,6 +75,28 @@ export interface IGovernance {
 export interface GnosisConfig {
   trustedAddresses: TrustedAddress[];
   signatureThreshold: string;
+}
+
+export enum VetoVotingType {
+  ERC20,
+  MULTISIG,
+  UNKNOWN,
+}
+
+export interface IGnosisVetoContract {
+  vetoGuardContract: VetoGuard | UsulVetoGuard | undefined;
+  vetoVotingContract: VetoERC20Voting | VetoMultisigVoting | undefined;
+  vetoVotingType: VetoVotingType;
+}
+
+export interface IGnosisFreezeData {
+  freezeVotesThreshold: BigNumber; // Number of freeze votes required to activate a freeze
+  freezeProposalCreatedTime: BigNumber; // Block number the freeze proposal was created at
+  freezeProposalVoteCount: BigNumber; // Number of accrued freeze votes
+  freezeProposalPeriod: BigNumber; // Number of blocks a freeze proposal has to succeed
+  freezePeriod: BigNumber; // Number of blocks a freeze lasts, from time of freeze proposal creation
+  userHasFreezeVoted: boolean;
+  isFrozen: boolean;
 }
 
 export interface GnosisDAO extends DAODetails, GnosisConfig {}
